fix klee skill brick on ready check
If Klee attempted to cast her skill twice in rapid succession, the sim
would immediately attempt to queue another skill even if there wouldn't
be one available. This is because the 33f delay on her skill cooldown
gave the sim enough time to determine skill ready. To fix this, we
change `CanQueueAfter` to a higher value.