add cyno burst pp slide
If cyno performs an attack, charge, or skill, then the burst modifier
will be extended by 1s every 0.5s for 1.5s. This is a convoluted way to
implement a pp slide, so instead we're extending the burst only if we
perform one of these actions as the buff is about to expire.