rework candace q to use confer system
Candace's Q has two modifier types: grandfather and local. The
grandfather is responsible for distributing the local modifier for the
duration of the burst. The local modifier is the actual dmg%, a4, wave,
etc effects.

There's not an easy way to implement this right now, so the work around
is to use expiry=-1 AttackMods for representing the "local" modifier.
Once the "grandfather" candace-q is added to candace herself, we enable
the buffs as expected.